♻️ Refactor submission queries to own repository file
